pyproject.toml
[project]
name = "api-explorer"
version = "0.2.0"
description = "Lightweight API testing tool."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "PyYAML>=6.0.2",
  "urllib3>=2.6.3",
  "idna>=3.11"
]

[project.scripts]
api-cli = "api_cli:main"

[dependency-groups]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.11.0",
  "ty>=0.0.1a17",
]

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["api_cli", "api_client"]
